feat(argument): @defaultIfUndefined decorator
This PR adds a `@defaultIfUndefined` decorator which modifies
@argument's behavior. Normally, it will only initialize its value if
the key does not yet exist on the object - this allows users to pass
any value, including `undefined`, to objects and components and have it
be set correctly. Adding this decorator will cause the default value to
override the passed in value if the passed in value is `undefined`.

This is especially useful when creating wrapper components, where the
user may wish to pass through certain arguments to a sub-component:

```
{{sub-component someArgument=someArgument}}
```

With just the `@argument` decorator, the user would have to provide a
default value for `someArgument` in their component, otherwise it would
be `undefined`.