bugfix(validators): Make validators work with computeds
Currently validators do not work with computeds. The underlying value will
be gotten the first time and cached forever, throwing away any computed
value. This change makes validators check to see if the cached value is a
computed property, and if so proxy directly to it. In the future we can
check to see if the value has an ES getter as well.