feat(required): Allow required fields to be provided by class
Currently required fields are only accepted if they are on the instance
(hasOwnProperty). This is unflexible, as a subclass may choose to provide
an argument via a computed or getter which would be valid, but would not
meet this requirement. The additional check verifies that the field is
empty, and then we check to see if it also has not been defined on the
instance (undefined only makes sense as an instance value).