refactor(argument): Use get/set instead of initializers (#52)
This PR refactors the internals of @argument to make it use getters
and setters on the prototype to provide default values instead of storing
initializers. This is more inline with previous versions of Ember, and in
fact makes the logic for defaultIfUndefined and defaultIfNullish even easier
to implement.

This will likely not be necessary for future versions of @argument, but is
likely the best strategy for backwards compatibility until then.