fix `optional` helper (#71)
Basically the problem I was facing was that using, for example, `@type(optional(shapeOf({ then: Function })))` was throwing if I passed in `null or undefined`, which shouldn't because it was marked as optional.

I fixed this by testing null and undefined first and only then testing the actual type that is passed to `optional`. This change made the test pass.

Let me know if you need anything else from me.