fix: actually run component extension in tests
The new `@argument` behavior only injects the Component behavior
extension into classes that have an argument defined, instead of
adding it to _all_ components. However, many test cases didn't actually
define an argument, meaning the injecting never took place. Thus, the
tests aren't actually validating the behavior that they should have
validated.

This defined an optional argument on all the test classes, so that
the injection takes place in each of them and the behavior is actually
tested.